+++ azalea-client/src/plugins/respawn.rs
@@ -2,7 +2,7 @@ use azalea_protocol::packets::game::s_client_command::{self, ServerboundClientCo
 use bevy_app::{App, Plugin, Update};
 use bevy_ecs::prelude::*;
 
-use crate::packet::game::SendPacketEvent;
+use crate::packet::game::SendGamePacketEvent;
 
 /// Tell the server that we're respawning.
 #[derive(Message, Debug, Clone)]
@@ -21,7 +21,7 @@ impl Plugin for RespawnPlugin {
 
 pub fn perform_respawn(mut events: MessageReader<PerformRespawnEvent>, mut commands: Commands) {
     for event in events.read() {
-        commands.trigger(SendPacketEvent::new(
+        commands.trigger(SendGamePacketEvent::new(
             event.entity,
             ServerboundClientCommand {
                 action: s_client_command::Action::PerformRespawn,